Finished bridge package (for now)

+++ bridge/tests/ClientServerTest.php
@@ -2,20 +2,21 @@
 
 declare(strict_types=1);
 
-namespace MinVWS\DUSi\Shared\Bridge\Tests\Client;
+namespace MinVWS\DUSi\Shared\Bridge\Tests;
 
+use Exception;
 use MinVWS\DUSi\Shared\Bridge\Client\Client;
 use MinVWS\DUSi\Shared\Bridge\Client\Exceptions\TimeoutException;
 use MinVWS\DUSi\Shared\Bridge\Server\Server;
-use MinVWS\DUSi\Shared\Bridge\Tests\EchoParams;
-use MinVWS\DUSi\Shared\Bridge\Tests\EchoResult;
-use PhpAmqpLib\Connection\AMQPStreamConnection;
+use MinVWS\DUSi\Shared\Bridge\Shared\Connection;
+use MinVWS\DUSi\Shared\Bridge\Tests\DTO\EchoParams;
+use MinVWS\DUSi\Shared\Bridge\Tests\DTO\EchoResult;
 use PHPUnit\Framework\TestCase;
 
 use function pcntl_fork;
 use function posix_kill;
 
-class ClientTest extends TestCase
+class ClientServerTest extends TestCase
 {
     private ?int $serverPid;
 
@@ -29,10 +30,10 @@ class ClientTest extends TestCase
         }
 
         // child process, start server and wait indefinitely
-        $conn = new AMQPStreamConnection('rabbitmq', 5672, 'guest', 'guest');
+        $conn = Connection::create(host: 'rabbitmq');
         $server = new Server($conn);
 
-        $conn->channel()->queue_purge('rpc_queue');
+        $conn->connection->channel()->queue_purge('rpc_queue', true);
 
         $server->bind('echo', EchoParams::class, static function (EchoParams $params) {
             return new EchoResult(str_repeat($params->message, $params->times));
@@ -45,22 +46,31 @@ class ClientTest extends TestCase
     protected function tearDown(): void
     {
         parent::tearDown();
-        $this->assertTrue(posix_kill($this->serverPid, SIGKILL));
+
+        if ($this->serverPid !== null) {
+            $this->assertTrue(posix_kill($this->serverPid, SIGKILL));
+        }
     }
 
+    /**
+     * @throws Exception
+     */
     public function testIfAMethodCanBeCalled(): void
     {
         parent::setUp();
 
-        $conn = new AMQPStreamConnection('rabbitmq', 5672, 'guest', 'guest');
+        $conn = Connection::create(host: 'rabbitmq');
         $client = new Client($conn);
         $result = $client->call('echo', new EchoParams('ha', 3), EchoResult::class, timeout: 1);
         $this->assertEquals('hahaha', $result->message);
     }
 
+    /**
+     * @throws Exception
+     */
     public function testIfANonExistentMethodResultsInATimeout(): void
     {
-        $conn = new AMQPStreamConnection('rabbitmq', 5672, 'guest', 'guest');
+        $conn = Connection::create(host: 'rabbitmq');
         $client = new Client($conn);
         $this->expectException(TimeoutException::class);
         $client->call('methodDoesNotExist', new EchoParams('ha', 3), EchoResult::class, timeout: 1);
